+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added support to import public accounts to `Client` (#733).
 * [BREAKING] Merged `TonicRpcClient` with `WebTonicRpcClient` and added missing endpoints (#744).
+* Added support for script execution in the `Client` and CLI (#777).
 
 ### Changes